Secrets: Fixed pet HP resetting if the pet is above 32k when zoning.
Secrets: Fixed player HP rollover bug (~495k hitpoints, rarely encountered) with AAs by temporarily casting it as a float.

Secrets: Fixed damage on weapons above 255. They only display on SoF+ clients, and Titanium will simply show a cosmetic issue (an incorrect value). You can set an item to any value up to the max for an uint32 and it will work.

Secrets: Added rules for the low-level damage cap. There is a rule for the pre-10 cap, and one for the pre-20 cap. Defaults to 20 and 40 respectively.

REQUIRED: You will have to recompile sharedmem with this update. Items will fail to load otherwise.

+++ trunk/EQEmuServer/common/shareddb.cpp
@@ -829,7 +829,7 @@ bool SharedDatabase::DBLoadItems(sint32 iItemCount, uint32 iMaxItemID) {
 			item.ElemDmgType = (uint8)atoi(row[ItemField::elemdmgtype]);
 			item.ElemDmgAmt = (uint8)atoi(row[ItemField::elemdmgamt]);
 			item.Range = (uint8)atoi(row[ItemField::range]);
-			item.Damage = (uint8)atoi(row[ItemField::damage]);
+			item.Damage = (uint32)atoi(row[ItemField::damage]);
 			item.Color = (uint32)atoul(row[ItemField::color]);
 			item.Classes = (uint32)atoul(row[ItemField::classes]);
 			item.Races = (uint32)atoul(row[ItemField::races]);
